chore: acelere os testes em test_localsearch

+++ tests/test_localsearch.py
@@ -6,37 +6,44 @@ import pytest
 import json
 import os
 from lunr.index import Index
-from .utils import without_elastic_search
+from .utils import without_elastic_search, without_artifact_auto_processing
+from unittest import mock
 
+@without_artifact_auto_processing()
 @without_elastic_search()
 @pytest.mark.last
 @pytest.mark.django_db(transaction=True)
-def test_localsearch_remote_elastic_lunr(settings):
+def test_localsearch_remote_elastic_lunr(settings, mocker):
     #FIXME: mocar os métodos de NewsFetcher
     settings.LUNR_INDEX_BACKEND = 'remote'
     settings.LUNR_INDEX_REMOTE_HOST = 'http://localhost:5000'
     settings.LUNR_INDEX_REMOTE_SECRET = 'alabamba'
 
-    with requests_mock.Mocker() as m:
-        m.register_uri('POST', settings.LUNR_INDEX_REMOTE_HOST, text='success')
-        with basic_news() as news:
-            news.save()
-        # verifique se a requisição foi chamada
-        assert m.called
-        # verifique se foi enviada uma requisição para o servidor configurado
-        reqs = [r for r in m.request_history if settings.LUNR_INDEX_REMOTE_HOST in r.url]
-        assert len(reqs)
+    mocked_client = mocker.Mock()
+    mocker.patch('xram_memory.lunr_index.lib.index_builders.RemoteElasticLunrIndexBuilder._get_client', return_value=mocked_client)
+    mocker.patch('xram_memory.lunr_index.signals.celery_is_avaliable', return_value=False)
 
+
+    with basic_news() as news:
+        news.save()
+        # verifique se a requisição foi chamada
+        assert mocked_client.post.called
+        mocked_client.post.called_with(
+            settings.LUNR_INDEX_REMOTE_HOST,
+            headers={'Authorization': f'Bearer {settings.LUNR_INDEX_REMOTE_SECRET}'}
+        )
+
+@without_artifact_auto_processing()
 @without_elastic_search()
 @pytest.mark.last
 @pytest.mark.django_db(transaction=True)
-def test_localsearch_lunr_py(settings):
+def test_localsearch_lunr_py(settings, mocker):
     #FIXME: mocar os métodos de NewsFetcher
     settings.LUNR_INDEX_BACKEND = 'local'
-
     try:
         fd, file_path, = tempfile.mkstemp(dir=settings.MEDIA_ROOT)
         settings.LUNR_INDEX_FILE_PATH = file_path
+        mocker.patch('xram_memory.lunr_index.signals.celery_is_avaliable', return_value=False)
         with basic_news() as news:
             news.save()
             with open(file_path, 'r') as json_index:
@@ -47,4 +54,5 @@ def test_localsearch_lunr_py(settings):
         try:
             os.close(fd)
         finally:
-            os.remove(file_path)
+            if (file_path):
+                os.remove(file_path)
